requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
groq
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.1